[BACKEND] [REFACTOR] Added s3.utils.ts file for small utilitary functions
